Emergency Calls for Montgomery County, PA

The emergency calls coming in at Montgomery County, PA since 2015-12-10. Data set was created at 2016-10-17 16:15:40 CEST from the API and contains events until 2016-10-17 09:47:03 EST. From the original set the columns desc and e are not included.

emergency

Format

A data frame with 120450 rows and 6 variables:

lat

Latitude from Google maps, based on the address

lng

Longitude from Google maps, based on the address

zip

Zipcode from Google, when possible

title

Title, emergency category

time_stamp

YYYY-MM-DD HH:MM:SS

twp

Township

Source

https://storage.googleapis.com/montco-stats/tzr.csv